Fix all critical bugs found in pipeline audit
scope_pre (critical): off-by-one — pixel 128 (scene mean) was dead code
  because float comparison `pos.x <= float(SCOPE_BINS)` fails at pixel
  center 128.5; fixed to `int(pos.x) <= SCOPE_BINS`

creative_render_chain (critical): ZONE_LERP_SPEED/CHROMA_LERP_SPEED at
  0.01 gave ~167s settling time — zone medians never adapted after frame 0;
  all LERP speeds unified to 4.3 (0.7s settling, matches corrective chain)

creative_render_chain (critical): ApplyContrast used fixed strength with
  no IQR; ApplyChroma had no sat_w guard and no IQR; CreativeZoneLevelsTex
  and CreativeSatLevelsTex were R16F (median only); all brought to parity
  with corrective chain: RGBA16F, p25/p50/p75 CDF walk, IQR-adaptive
  strength, sat_w = smoothstep(0, 0.15, hsv.y) near-neutral guard

analysis_frame_analysis: histogram smooth EMA not frametime-normalized;
  added frametime uniform, fixed both LumHistSmooth and SatHistSmooth;
  LERP_SPEED 8 → 4.3

analysis_scope: post-correction mean EMA not frametime-normalized; added
  frametime uniform, fixed SCOPE_LERP formula; SCOPE_LERP 3 → 4.3

+++ general/analysis-scope/analysis_scope_pre.fx
@@ -40,7 +40,7 @@ float Luma(float3 c) { return dot(c, float3(0.2126, 0.7152, 0.0722)); }
 float4 ScopeCapturePS(float4 pos : SV_Position,
                       float2 uv  : TEXCOORD0) : SV_Target
 {
-    if (pos.y < 1.0 && pos.x <= float(SCOPE_BINS))
+    if (pos.y < 1.0 && int(pos.x) <= SCOPE_BINS)
     {
         // Sample grid once — used for both histogram and mean
         float samples[256];
